@@ -2,22 +2,46 @@
 """
 ChatterBot setup file.
 """
+import os
+import sys
+import platform
+import configparser
 from setuptools import setup
 
 
-# Dynamically retrieve the version information from the chatterbot module
-CHATTERBOT = __import__('chatterbot')
-VERSION = CHATTERBOT.__version__
-AUTHOR = CHATTERBOT.__author__
-AUTHOR_EMAIL = CHATTERBOT.__email__
-URL = CHATTERBOT.__url__
-DESCRIPTION = CHATTERBOT.__doc__
+if sys.version_info[0] < 3:
+    raise Exception(
+        'You are tying to install ChatterBot on Python version {}.\n'
+        'Please install ChatterBot in Python 3 instead.'.format(
+            platform.python_version()
+        )
+    )
+
+config = configparser.ConfigParser()
+
+current_directory = os.path.dirname(os.path.abspath(__file__))
+config_file_path = os.path.join(current_directory, 'setup.cfg')
+
+config.read(config_file_path)
+
+VERSION = config['chatterbot']['version']
+AUTHOR = config['chatterbot']['author']
+AUTHOR_EMAIL = config['chatterbot']['email']
+URL = config['chatterbot']['url']
 
 with open('README.md') as f:
     LONG_DESCRIPTION = f.read()
 
+REQUIREMENTS = []
+DEPENDENCIES = []
+
 with open('requirements.txt') as requirements:
-    REQUIREMENTS = requirements.readlines()
+    for requirement in requirements.readlines():
+        if requirement.startswith('git+git://'):
+            DEPENDENCIES.append(requirement)
+        else:
+            REQUIREMENTS.append(requirement)
+
 
 setup(
     name='ChatterBot',
@@ -27,18 +51,15 @@ setup(
     project_urls={
         'Documentation': 'https://chatterbot.readthedocs.io',
     },
-    description=DESCRIPTION,
+    description='ChatterBot is a machine learning, conversational dialog engine.',
     long_description=LONG_DESCRIPTION,
     long_description_content_type='text/markdown',
     author=AUTHOR,
     author_email=AUTHOR_EMAIL,
     packages=[
         'chatterbot',
-        'chatterbot.input',
-        'chatterbot.output',
         'chatterbot.storage',
         'chatterbot.logic',
-        'chatterbot.api',
         'chatterbot.ext',
         'chatterbot.ext.sqlalchemy_app',
         'chatterbot.ext.django_chatterbot',
@@ -47,7 +68,8 @@ setup(
     package_dir={'chatterbot': 'chatterbot'},
     include_package_data=True,
     install_requires=REQUIREMENTS,
-    python_requires='>=2.7, <4',
+    dependency_links=DEPENDENCIES,
+    python_requires='>=3.4, <=3.8',
     license='BSD',
     zip_safe=True,
     platforms=['any'],
@@ -68,6 +90,7 @@ setup(
         'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
+        'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3 :: Only',
     ],
     test_suite='tests'
