@@ -18,11 +18,11 @@ class MongoDatabaseAdapter(StorageAdapter):
     """
 
     def __init__(self, **kwargs):
-        super(MongoDatabaseAdapter, self).__init__(**kwargs)
+        super().__init__(**kwargs)
         from pymongo import MongoClient
         from pymongo.errors import OperationFailure
 
-        self.database_uri = self.kwargs.get(
+        self.database_uri = kwargs.get(
             'database_uri', 'mongodb://localhost:27017/chatterbot-database'
         )
 
@@ -80,6 +80,7 @@ class MongoDatabaseAdapter(StorageAdapter):
         exclude_text = kwargs.pop('exclude_text', None)
         exclude_text_words = kwargs.pop('exclude_text_words', [])
         persona_not_startswith = kwargs.pop('persona_not_startswith', None)
+        search_text_contains = kwargs.pop('search_text_contains', None)
 
         if tags:
             kwargs['tags'] = {
@@ -119,6 +120,12 @@ class MongoDatabaseAdapter(StorageAdapter):
                 }
             kwargs['persona']['$not'] = re.compile('^bot:*')
 
+        if search_text_contains:
+            or_regex = '|'.join([
+                '{}'.format(re.escape(word)) for word in search_text_contains.split(' ')
+            ])
+            kwargs['search_text'] = re.compile(or_regex)
+
         mongo_ordering = []
 
         if order_by:
@@ -152,11 +159,11 @@ class MongoDatabaseAdapter(StorageAdapter):
             kwargs['tags'] = list(set(kwargs['tags']))
 
         if 'search_text' not in kwargs:
-            kwargs['search_text'] = self.stemmer.get_bigram_pair_string(kwargs['text'])
+            kwargs['search_text'] = self.tagger.get_text_index_string(kwargs['text'])
 
         if 'search_in_response_to' not in kwargs:
             if kwargs.get('in_response_to'):
-                kwargs['search_in_response_to'] = self.stemmer.get_bigram_pair_string(kwargs['in_response_to'])
+                kwargs['search_in_response_to'] = self.tagger.get_text_index_string(kwargs['in_response_to'])
 
         inserted = self.statements.insert_one(kwargs)
 
@@ -171,22 +178,15 @@ class MongoDatabaseAdapter(StorageAdapter):
         create_statements = []
 
         for statement in statements:
-            statement_data = {
-                'text': statement.text,
-                'search_text': statement.search_text,
-                'conversation': statement.conversation,
-                'persona': statement.persona,
-                'in_response_to': statement.in_response_to,
-                'search_in_response_to': statement.search_in_response_to,
-                'created_at': statement.created_at,
-                'tags': list(set(statement.tags))
-            }
+            statement_data = statement.serialize()
+            tag_data = list(set(statement_data.pop('tags', [])))
+            statement_data['tags'] = tag_data
 
             if not statement.search_text:
-                statement_data['search_text'] = self.stemmer.get_bigram_pair_string(statement.text)
+                statement_data['search_text'] = self.tagger.get_text_index_string(statement.text)
 
             if not statement.search_in_response_to and statement.in_response_to:
-                statement_data['search_in_response_to'] = self.stemmer.get_bigram_pair_string(statement.in_response_to)
+                statement_data['search_in_response_to'] = self.tagger.get_text_index_string(statement.in_response_to)
 
             create_statements.append(statement_data)
 
@@ -197,10 +197,10 @@ class MongoDatabaseAdapter(StorageAdapter):
         data.pop('id', None)
         data.pop('tags', None)
 
-        data['search_text'] = self.stemmer.get_bigram_pair_string(data['text'])
+        data['search_text'] = self.tagger.get_text_index_string(data['text'])
 
         if data.get('in_response_to'):
-            data['search_in_response_to'] = self.stemmer.get_bigram_pair_string(data['in_response_to'])
+            data['search_in_response_to'] = self.tagger.get_text_index_string(data['in_response_to'])
 
         update_data = {
             '$set': data
