@@ -2,15 +2,15 @@
 This module contains various text-comparison algorithms
 designed to compare one statement to another.
 """
-
-# Use python-Levenshtein if available
-try:
-    from Levenshtein.StringMatcher import StringMatcher as SequenceMatcher
-except ImportError:
-    from difflib import SequenceMatcher
+from chatterbot.exceptions import OptionalDependencyImportError
+from difflib import SequenceMatcher
 
 
 class Comparator:
+
+    def __init__(self, language):
+
+        self.language = language
 
     def __call__(self, statement_a, statement_b):
         return self.compare(statement_a, statement_b)
@@ -29,7 +29,7 @@ class LevenshteinDistance(Comparator):
     based on the Levenshtein distance algorithm.
     """
 
-    def compare(self, statement, other_statement):
+    def compare(self, statement_a, statement_b):
         """
         Compare the two input statements.
 
@@ -38,17 +38,17 @@ class LevenshteinDistance(Comparator):
         """
 
         # Return 0 if either statement has a falsy text value
-        if not statement.text or not other_statement.text:
+        if not statement_a.text or not statement_b.text:
             return 0
 
         # Get the lowercase version of both strings
-        statement_text = str(statement.text.lower())
-        other_statement_text = str(other_statement.text.lower())
+        statement_a_text = str(statement_a.text.lower())
+        statement_b_text = str(statement_b.text.lower())
 
         similarity = SequenceMatcher(
             None,
-            statement_text,
-            other_statement_text
+            statement_a_text,
+            statement_b_text
         )
 
         # Calculate a decimal percent of the similarity
@@ -57,148 +57,36 @@ class LevenshteinDistance(Comparator):
         return percent
 
 
-class SynsetDistance(Comparator):
+class SpacySimilarity(Comparator):
     """
-    Calculate the similarity of two statements.
-    This is based on the total maximum synset similarity between each word in each sentence.
-
-    This algorithm uses the `wordnet`_ functionality of `NLTK`_ to determine the similarity
-    of two statements based on the path similarity between each token of each statement.
-    This is essentially an evaluation of the closeness of synonyms.
+    Calculate the similarity of two statements using Spacy models.
     """
 
-    def initialize_nltk_wordnet(self):
-        """
-        Download required NLTK corpora if they have not already been downloaded.
-        """
-        from chatterbot.utils import nltk_download_corpus
+    def __init__(self, language):
+        super().__init__(language)
+        try:
+            import spacy
+        except ImportError:
+            message = (
+                'Unable to import "spacy".\n'
+                'Please install "spacy" before using the SpacySimilarity comparator:\n'
+                'pip3 install "spacy>=2.1,<2.2"'
+            )
+            raise OptionalDependencyImportError(message)
 
-        nltk_download_corpus('corpora/wordnet')
+        self.nlp = spacy.load(self.language.ISO_639_1)
 
-    def initialize_nltk_punkt(self):
-        """
-        Download required NLTK corpora if they have not already been downloaded.
-        """
-        from chatterbot.utils import nltk_download_corpus
-
-        nltk_download_corpus('tokenizers/punkt')
-
-    def initialize_nltk_stopwords(self):
-        """
-        Download required NLTK corpora if they have not already been downloaded.
-        """
-        from chatterbot.utils import nltk_download_corpus
-
-        nltk_download_corpus('corpora/stopwords')
-
-    def compare(self, statement, other_statement):
+    def compare(self, statement_a, statement_b):
         """
         Compare the two input statements.
 
         :return: The percent of similarity between the closest synset distance.
         :rtype: float
-
-        .. _wordnet: http://www.nltk.org/howto/wordnet.html
-        .. _NLTK: http://www.nltk.org/
         """
-        from nltk.corpus import wordnet
-        from nltk import word_tokenize
-        from chatterbot import utils
-        import itertools
+        document_a = self.nlp(statement_a.text)
+        document_b = self.nlp(statement_b.text)
 
-        tokens1 = word_tokenize(statement.text.lower())
-        tokens2 = word_tokenize(other_statement.text.lower())
-
-        # Remove all stop words from the list of word tokens
-        tokens1 = utils.remove_stopwords(tokens1, language='english')
-        tokens2 = utils.remove_stopwords(tokens2, language='english')
-
-        # The maximum possible similarity is an exact match
-        # Because path_similarity returns a value between 0 and 1,
-        # max_possible_similarity is the number of words in the longer
-        # of the two input statements.
-        max_possible_similarity = min(
-            len(statement.text.split()),
-            len(other_statement.text.split())
-        ) / max(
-            len(statement.text.split()),
-            len(other_statement.text.split())
-        )
-
-        max_similarity = 0.0
-
-        # Get the highest matching value for each possible combination of words
-        for combination in itertools.product(*[tokens1, tokens2]):
-
-            synset1 = wordnet.synsets(combination[0])
-            synset2 = wordnet.synsets(combination[1])
-
-            if synset1 and synset2:
-
-                # Get the highest similarity for each combination of synsets
-                for synset in itertools.product(*[synset1, synset2]):
-                    similarity = synset[0].path_similarity(synset[1])
-
-                    if similarity and (similarity > max_similarity):
-                        max_similarity = similarity
-
-        if max_possible_similarity == 0:
-            return 0
-
-        return max_similarity / max_possible_similarity
-
-
-class SentimentComparison(Comparator):
-    """
-    Calculate the similarity of two statements based on the closeness of
-    the sentiment value calculated for each statement.
-    """
-
-    def initialize_nltk_vader_lexicon(self):
-        """
-        Download the NLTK vader lexicon for sentiment analysis
-        that is required for this algorithm to run.
-        """
-        from chatterbot.utils import nltk_download_corpus
-
-        nltk_download_corpus('sentiment/vader_lexicon')
-
-    def compare(self, statement, other_statement):
-        """
-        Return the similarity of two statements based on
-        their calculated sentiment values.
-
-        :return: The percent of similarity between the sentiment value.
-        :rtype: float
-        """
-        from nltk.sentiment.vader import SentimentIntensityAnalyzer
-
-        sentiment_analyzer = SentimentIntensityAnalyzer()
-        statement_polarity = sentiment_analyzer.polarity_scores(statement.text.lower())
-        statement2_polarity = sentiment_analyzer.polarity_scores(other_statement.text.lower())
-
-        statement_greatest_polarity = 'neu'
-        statement_greatest_score = -1
-        for polarity in sorted(statement_polarity):
-            if statement_polarity[polarity] > statement_greatest_score:
-                statement_greatest_polarity = polarity
-                statement_greatest_score = statement_polarity[polarity]
-
-        statement2_greatest_polarity = 'neu'
-        statement2_greatest_score = -1
-        for polarity in sorted(statement2_polarity):
-            if statement2_polarity[polarity] > statement2_greatest_score:
-                statement2_greatest_polarity = polarity
-                statement2_greatest_score = statement2_polarity[polarity]
-
-        # Check if the polarity if of a different type
-        if statement_greatest_polarity != statement2_greatest_polarity:
-            return 0
-
-        values = [statement_greatest_score, statement2_greatest_score]
-        difference = max(values) - min(values)
-
-        return 1.0 - difference
+        return document_a.similarity(document_b)
 
 
 class JaccardSimilarity(Comparator):
@@ -227,93 +115,39 @@ class JaccardSimilarity(Comparator):
     .. _`Jaccard similarity index`: https://en.wikipedia.org/wiki/Jaccard_index
     """
 
-    def initialize_nltk_wordnet(self):
-        """
-        Download the NLTK wordnet corpora that is required for this algorithm
-        to run only if the corpora has not already been downloaded.
-        """
-        from chatterbot.utils import nltk_download_corpus
+    def __init__(self, language):
+        super().__init__(language)
+        try:
+            import spacy
+        except ImportError:
+            message = (
+                'Unable to import "spacy".\n'
+                'Please install "spacy" before using the JaccardSimilarity comparator:\n'
+                'pip3 install "spacy>=2.1,<2.2"'
+            )
+            raise OptionalDependencyImportError(message)
 
-        nltk_download_corpus('corpora/wordnet')
+        self.nlp = spacy.load(self.language.ISO_639_1)
 
-    def initialize_nltk_averaged_perceptron_tagger(self):
-        """
-        Download the NLTK averaged perceptron tagger that is required for this algorithm
-        to run only if the corpora has not already been downloaded.
-        """
-        from chatterbot.utils import nltk_download_corpus
-
-        nltk_download_corpus('averaged_perceptron_tagger')
-
-    def initialize_nltk_stopwords(self):
-        """
-        Download required NLTK corpora if they have not already been downloaded.
-        """
-        from chatterbot.utils import nltk_download_corpus
-
-        nltk_download_corpus('corpora/stopwords')
-
-    def compare(self, statement, other_statement):
+    def compare(self, statement_a, statement_b):
         """
         Return the calculated similarity of two
         statements based on the Jaccard index.
         """
-        from nltk.corpus import wordnet
-        import nltk
-        import string
-
-        # Get default English stopwords
-        stopwords = nltk.corpus.stopwords.words('english')
-
-        lemmatizer = nltk.stem.wordnet.WordNetLemmatizer()
-
         # Make both strings lowercase
-        a = statement.text.lower()
-        b = other_statement.text.lower()
+        document_a = self.nlp(statement_a.text.lower())
+        document_b = self.nlp(statement_b.text.lower())
 
-        # Remove punctuation from each string
-        table = str.maketrans(dict.fromkeys(string.punctuation))
-        a = a.translate(table)
-        b = b.translate(table)
-
-        def get_wordnet_pos(pos_tag):
-            if pos_tag[1].startswith('J'):
-                return (pos_tag[0], wordnet.ADJ)
-            elif pos_tag[1].startswith('V'):
-                return (pos_tag[0], wordnet.VERB)
-            elif pos_tag[1].startswith('N'):
-                return (pos_tag[0], wordnet.NOUN)
-            elif pos_tag[1].startswith('R'):
-                return (pos_tag[0], wordnet.ADV)
-            else:
-                return (pos_tag[0], wordnet.NOUN)
-
-        pos_a = list(map(get_wordnet_pos, nltk.pos_tag(nltk.tokenize.word_tokenize(a))))
-        pos_b = list(map(get_wordnet_pos, nltk.pos_tag(nltk.tokenize.word_tokenize(b))))
-
-        lemma_a = [
-            lemmatizer.lemmatize(
-                token, pos
-            ) for token, pos in pos_a if token not in stopwords
-        ]
-        lemma_b = [
-            lemmatizer.lemmatize(
-                token, pos
-            ) for token, pos in pos_b if token not in stopwords
-        ]
+        statement_a_lemmas = set([
+            token.lemma_ for token in document_a if not token.is_stop
+        ])
+        statement_b_lemmas = set([
+            token.lemma_ for token in document_b if not token.is_stop
+        ])
 
         # Calculate Jaccard similarity
-        numerator = len(set(lemma_a).intersection(lemma_b))
-        denominator = float(len(set(lemma_a).union(lemma_b)))
+        numerator = len(statement_a_lemmas.intersection(statement_b_lemmas))
+        denominator = float(len(statement_a_lemmas.union(statement_b_lemmas)))
         ratio = numerator / denominator
 
         return ratio
-
-
-# ---------------------------------------- #
-
-
-levenshtein_distance = LevenshteinDistance()
-synset_distance = SynsetDistance()
-sentiment_comparison = SentimentComparison()
-jaccard_similarity = JaccardSimilarity()
