@@ -17,26 +17,6 @@ def import_module(dotted_path):
     return getattr(module, module_parts[-1])
 
 
-def get_initialization_functions(obj, attribute):
-    """
-    Return all initialization methods for the comparison algorithm.
-    Initialization methods must start with 'initialize_' and
-    take no parameters.
-    """
-    initialization_methods = {}
-
-    attribute_parts = attribute.split('.')
-    outermost_attribute = getattr(obj, attribute_parts.pop(0))
-    for next_attribute in attribute_parts:
-        outermost_attribute = getattr(outermost_attribute, next_attribute)
-
-    for method in dir(outermost_attribute):
-        if method.startswith('initialize_'):
-            initialization_methods[method] = getattr(outermost_attribute, method)
-
-    return initialization_methods
-
-
 def initialize_class(data, *args, **kwargs):
     """
     :param data: A string or dictionary containing a import_path attribute.
@@ -90,83 +70,6 @@ def validate_adapter_class(validate_class, adapter_class):
         )
 
 
-def nltk_download_corpus(resource_path):
-    """
-    Download the specified NLTK corpus file
-    unless it has already been downloaded.
-
-    Returns True if the corpus needed to be downloaded.
-    """
-    from nltk.data import find
-    from nltk import download
-    from os.path import split, sep
-    from zipfile import BadZipfile
-
-    # Download the NLTK data only if it is not already downloaded
-    _, corpus_name = split(resource_path)
-
-    # From http://www.nltk.org/api/nltk.html
-    # When using find() to locate a directory contained in a zipfile,
-    # the resource name must end with the forward slash character.
-    # Otherwise, find() will not locate the directory.
-    #
-    # Helps when resource_path=='sentiment/vader_lexicon''
-    if not resource_path.endswith(sep):
-        resource_path = resource_path + sep
-
-    downloaded = False
-
-    try:
-        find(resource_path)
-    except LookupError:
-        download(corpus_name)
-        downloaded = True
-    except BadZipfile:
-        raise BadZipfile(
-            'The NLTK corpus file being opened is not a zipfile, '
-            'or it has been corrupted and needs to be manually deleted.'
-        )
-
-    return downloaded
-
-
-def remove_stopwords(tokens, language):
-    """
-    Takes a language (i.e. 'english'), and a set of word tokens.
-    Returns the tokenized text with any stopwords removed.
-    Stop words are words like "is, the, a, ..."
-
-    Be sure to download the required NLTK corpus before calling this function:
-    - from chatterbot.utils import nltk_download_corpus
-    - nltk_download_corpus('corpora/stopwords')
-    """
-    from nltk.corpus import stopwords
-
-    # Get the stopwords for the specified language
-    stop_words = stopwords.words(language)
-
-    # Remove the stop words from the set of word tokens
-    tokens = set(tokens) - set(stop_words)
-
-    return tokens
-
-
-def get_greatest_confidence(statement, options):
-    """
-    Returns the greatest confidence value for a statement that occurs
-    multiple times in the set of options.
-
-    :param statement: A statement object.
-    :param options: A tuple in the format of (confidence, statement).
-    """
-    values = []
-    for option in options:
-        if option[1] == statement:
-            values.append(option[0])
-
-    return max(values)
-
-
 def get_response_time(chatbot, statement='Hello'):
     """
     Returns the amount of time taken for a given
@@ -210,7 +113,7 @@ def print_progress_bar(description, iteration_counter, total_items, progress_bar
     percent = float(iteration_counter) / total_items
     hashes = '#' * int(round(percent * progress_bar_length))
     spaces = ' ' * (progress_bar_length - len(hashes))
-    sys.stdout.write("\r{0}: [{1}] {2}%".format(description, hashes + spaces, int(round(percent * 100))))
+    sys.stdout.write('\r{0}: [{1}] {2}%'.format(description, hashes + spaces, int(round(percent * 100))))
     sys.stdout.flush()
     if total_items == iteration_counter:
-        print("\r")
+        print('\r')
