@@ -2,20 +2,20 @@ from chatterbot import ChatBot
 
 
 bot = ChatBot(
-    "Unit Converter",
+    'Unit Converter',
     logic_adapters=[
-        "chatterbot.logic.UnitConversion",
-    ],
-    input_adapter="chatterbot.input.InputAdapter",
-    output_adapter="chatterbot.output.OutputAdapter"
+        'chatterbot.logic.UnitConversion',
+    ]
 )
 
-questions = ['How many meters are in a kilometer?',
-             'How many meters are in one inch?',
-             '0 celsius to fahrenheit',
-             'one hour is how many minutes ?']
+questions = [
+    'How many meters are in a kilometer?',
+    'How many meters are in one inch?',
+    '0 celsius to fahrenheit',
+    'one hour is how many minutes ?'
+]
 
 # Prints the convertion given the specific question
-for q in questions:
-    response = bot.get_response(q)
-    print(q + " -  Response: " + response.text)
+for question in questions:
+    response = bot.get_response(question)
+    print(question + ' -  Response: ' + response.text)
