@@ -4,16 +4,13 @@ from chatterbot.trainers import ListTrainer
 
 # Create a new instance of a ChatBot
 bot = ChatBot(
-    'Default Response Example Bot',
+    'Example Bot',
     storage_adapter='chatterbot.storage.SQLStorageAdapter',
     logic_adapters=[
         {
-            'import_path': 'chatterbot.logic.BestMatch'
-        },
-        {
-            'import_path': 'chatterbot.logic.LowConfidenceAdapter',
-            'threshold': 0.65,
-            'default_response': 'I am sorry, but I do not understand.'
+            'import_path': 'chatterbot.logic.BestMatch',
+            'default_response': 'I am sorry, but I do not understand.',
+            'maximum_similarity_threshold': 0.90
         }
     ]
 )
