@@ -2,19 +2,17 @@ from chatterbot import ChatBot
 
 
 bot = ChatBot(
-    "Math & Time Bot",
+    'Math & Time Bot',
     logic_adapters=[
-        "chatterbot.logic.MathematicalEvaluation",
-        "chatterbot.logic.TimeLogicAdapter"
-    ],
-    input_adapter="chatterbot.input.InputAdapter",
-    output_adapter="chatterbot.output.OutputAdapter"
+        'chatterbot.logic.MathematicalEvaluation',
+        'chatterbot.logic.TimeLogicAdapter'
+    ]
 )
 
 # Print an example of getting one math based response
-response = bot.get_response("What is 4 + 9?")
+response = bot.get_response('What is 4 + 9?')
 print(response)
 
 # Print an example of getting one time based response
-response = bot.get_response("What time is it?")
+response = bot.get_response('What time is it?')
 print(response)
