@@ -7,26 +7,26 @@ from chatterbot import ChatBot
 
 # Create a new instance of a ChatBot
 bot = ChatBot(
-    "Terminal",
-    storage_adapter="chatterbot.storage.SQLStorageAdapter",
+    'Terminal',
+    storage_adapter='chatterbot.storage.SQLStorageAdapter',
     logic_adapters=[
-        "chatterbot.logic.MathematicalEvaluation",
-        "chatterbot.logic.TimeLogicAdapter",
-        "chatterbot.logic.BestMatch"
+        'chatterbot.logic.MathematicalEvaluation',
+        'chatterbot.logic.TimeLogicAdapter',
+        'chatterbot.logic.BestMatch'
     ],
-    input_adapter="chatterbot.input.TerminalAdapter",
-    output_adapter="chatterbot.output.TerminalAdapter",
-    database_uri="../database.db"
+    database_uri='sqlite:///database.sqlite3'
 )
 
-print("Type something to begin...")
+print('Type something to begin...')
 
 # The following loop will execute each time the user enters input
 while True:
     try:
-        # We pass a blank string to this method because the parameter
-        # is not used by the TerminalAdapter
-        bot_input = bot.get_response('')
+        user_input = input()
+
+        bot_response = bot.get_response(user_input)
+
+        print(bot_response)
 
     # Press ctrl-c or ctrl-d on the keyboard to exit
     except (KeyboardInterrupt, EOFError, SystemExit):
