@@ -5,55 +5,37 @@ from chatterbot.conversation import Statement
 
 class ChatterBotResponseTestCase(ChatBotTestCase):
 
-    def test_get_initialization_functions(self):
+    def test_conversation_values_persisted_to_response(self):
+        response = self.chatbot.get_response('Hello', persist_values_to_response={
+            'conversation': 'test 1'
+        })
+        self.assertEqual(response.conversation, 'test 1')
+
+    def test_tag_values_persisted_to_response(self):
+        response = self.chatbot.get_response('Hello', persist_values_to_response={
+            'tags': [
+                'tag 1',
+                'tag 2'
+            ]
+        })
+        self.assertEqual(len(response.tags), 2)
+        self.assertIn('tag 1', response.get_tags())
+        self.assertIn('tag 2', response.get_tags())
+
+    def test_in_response_to_provided(self):
         """
-        Test that the initialization functions are returned.
+        Test that the process of looking up the previous response
+        in the conversation is ignored if a previous response is provided.
         """
-        functions = self.chatbot.get_initialization_functions()
-
-        self.assertIn('initialize_nltk_stopwords', functions)
-        self.assertIsLength(functions, 1)
-
-    def test_get_initialization_functions_synset_distance(self):
-        """
-        Test that the initialization functions are returned.
-        """
-        from chatterbot.comparisons import synset_distance
-
-        self.chatbot.logic_adapters[0].compare_statements = synset_distance
-        functions = self.chatbot.get_initialization_functions()
-
-        self.assertIn('initialize_nltk_stopwords', functions)
-        self.assertIn('initialize_nltk_wordnet', functions)
-        self.assertIn('initialize_nltk_punkt', functions)
-        self.assertIsLength(functions, 3)
-
-    def test_get_initialization_functions_sentiment_comparison(self):
-        """
-        Test that the initialization functions are returned.
-        """
-        from chatterbot.comparisons import sentiment_comparison
-
-        self.chatbot.logic_adapters[0].compare_statements = sentiment_comparison
-        functions = self.chatbot.get_initialization_functions()
-
-        self.assertIn('initialize_nltk_stopwords', functions)
-        self.assertIn('initialize_nltk_vader_lexicon', functions)
-        self.assertIsLength(functions, 2)
-
-    def test_get_initialization_functions_jaccard_similarity(self):
-        """
-        Test that the initialization functions are returned.
-        """
-        from chatterbot.comparisons import jaccard_similarity
-
-        self.chatbot.logic_adapters[0].compare_statements = jaccard_similarity
-        functions = self.chatbot.get_initialization_functions()
-
-        self.assertIn('initialize_nltk_wordnet', functions)
-        self.assertIn('initialize_nltk_stopwords', functions)
-        self.assertIn('initialize_nltk_averaged_perceptron_tagger', functions)
-        self.assertIsLength(functions, 3)
+        self.chatbot.get_response(
+            text='Hello',
+            in_response_to='Unique previous response.'
+        )
+        statement = self.chatbot.storage.filter(
+            text='Hello',
+            in_response_to='Unique previous response.'
+        )
+        self.assertIsNotNone(statement)
 
     def test_no_statements_known(self):
         """
@@ -65,11 +47,12 @@ class ChatterBotResponseTestCase(ChatBotTestCase):
         results = list(self.chatbot.storage.filter(text=statement_text))
 
         self.assertEqual(response.text, statement_text)
-        self.assertEqual(response.confidence, 1)
+        self.assertEqual(response.confidence, 0)
 
-        # Make sure that the input was saved
-        self.assertIsLength(results, 1)
+        # Make sure that the input and output were saved
+        self.assertIsLength(results, 2)
         self.assertEqual(results[0].text, statement_text)
+        self.assertEqual(results[1].text, statement_text)
 
     def test_one_statement_known_no_response(self):
         """
@@ -138,8 +121,14 @@ class ChatterBotResponseTestCase(ChatBotTestCase):
         self.chatbot.storage.create(text='Hi', in_response_to=None)
         self.chatbot.storage.create(text='Hello', in_response_to='Hi')
 
-        first_response = self.chatbot.get_response('Hi')
-        second_response = self.chatbot.get_response('How are you?')
+        first_response = self.chatbot.get_response(
+            text='Hi',
+            conversation='test'
+        )
+        second_response = self.chatbot.get_response(
+            text='How are you?',
+            conversation='test'
+        )
 
         results = list(self.chatbot.storage.filter(text='How are you?'))
 
@@ -148,7 +137,7 @@ class ChatterBotResponseTestCase(ChatBotTestCase):
         self.assertEqual(first_response.in_response_to, 'Hi')
 
         self.assertEqual(second_response.confidence, 0)
-        self.assertEqual(second_response.in_response_to, 'Hi')
+        self.assertEqual(second_response.in_response_to, 'How are you?')
 
         # Make sure that the second response was saved to the database
         self.assertIsLength(results, 1)
@@ -161,8 +150,24 @@ class ChatterBotResponseTestCase(ChatBotTestCase):
         statement = Statement(text='Wow!', conversation='test')
         response = self.chatbot.get_response(statement)
 
-        self.assertEqual(statement.text, response)
+        self.assertEqual(statement.text, response.text)
         self.assertEqual(response.conversation, 'test')
+
+    def test_get_response_additional_response_selection_parameters(self):
+        self.chatbot.storage.create_many([
+            Statement('A', conversation='test_1'),
+            Statement('B', conversation='test_1', in_response_to='A'),
+            Statement('A', conversation='test_2'),
+            Statement('C', conversation='test_2', in_response_to='A'),
+        ])
+
+        statement = Statement(text='A', conversation='test_3')
+        response = self.chatbot.get_response(statement, additional_response_selection_parameters={
+            'conversation': 'test_2'
+        })
+
+        self.assertEqual(response.text, 'C')
+        self.assertEqual(response.conversation, 'test_3')
 
     def test_get_response_unicode(self):
         """
@@ -212,16 +217,18 @@ class ChatterBotResponseTestCase(ChatBotTestCase):
 
         results = list(self.chatbot.storage.filter(text='Hello'))
 
-        self.assertIsLength(results, 1)
+        self.assertIsLength(results, 2)
         self.assertIn('test', results[0].get_tags())
+        self.assertEqual(results[1].get_tags(), [])
 
     def test_get_response_with_text_and_kwargs(self):
         self.chatbot.get_response('Hello', conversation='greetings')
 
         results = list(self.chatbot.storage.filter(text='Hello'))
 
-        self.assertIsLength(results, 1)
+        self.assertIsLength(results, 2)
         self.assertEqual(results[0].conversation, 'greetings')
+        self.assertEqual(results[1].conversation, 'greetings')
 
     def test_get_response_missing_text(self):
         with self.assertRaises(self.chatbot.ChatBotException):
@@ -235,8 +242,8 @@ class ChatterBotResponseTestCase(ChatBotTestCase):
         statement = Statement(text='Many insects adopt a tripedal gait for rapid yet stable walking.')
         response = self.chatbot.generate_response(statement)
 
-        self.assertEqual(response, statement)
-        self.assertEqual(response.confidence, 1)
+        self.assertEqual(response.text, statement.text)
+        self.assertEqual(response.confidence, 0)
 
     def test_learn_response(self):
         previous_response = Statement(text='Define Hemoglobin.')
@@ -301,19 +308,18 @@ class ChatterBotResponseTestCase(ChatBotTestCase):
         ])
 
         results = list(self.chatbot.storage.filter(
-            search_text=self.chatbot.storage.stemmer.get_bigram_pair_string(
+            search_text=self.chatbot.storage.tagger.get_text_index_string(
                 'Example A for search.'
             )
         ))
 
+        self.assertEqual(len(results), 1)
         self.assertEqual('Example A for search.', results[0].text)
-        self.assertEqual('Example B for search.', results[1].text)
-        self.assertIsLength(results, 2)
 
 
 class TestAdapterA(LogicAdapter):
 
-    def process(self, statement):
+    def process(self, statement, additional_response_selection_parameters=None):
         response = Statement(text='Good morning.')
         response.confidence = 0.2
         return response
@@ -321,7 +327,7 @@ class TestAdapterA(LogicAdapter):
 
 class TestAdapterB(LogicAdapter):
 
-    def process(self, statement):
+    def process(self, statement, additional_response_selection_parameters=None):
         response = Statement(text='Good morning.')
         response.confidence = 0.5
         return response
@@ -329,7 +335,7 @@ class TestAdapterB(LogicAdapter):
 
 class TestAdapterC(LogicAdapter):
 
-    def process(self, statement):
+    def process(self, statement, additional_response_selection_parameters=None):
         response = Statement(text='Good night.')
         response.confidence = 0.7
         return response
@@ -352,27 +358,13 @@ class ChatBotLogicAdapterTestCase(ChatBotTestCase):
         statement = self.chatbot.generate_response(Statement(text='Howdy!'))
 
         self.assertEqual(statement.confidence, 0.5)
-        self.assertEqual(statement, 'Good morning.')
-
-    def test_get_logic_adapters(self):
-        """
-        Test that all system logic adapters and regular logic adapters
-        can be retrieved as a list by a single method.
-        """
-        adapter_a = TestAdapterA(self.chatbot)
-        adapter_b = TestAdapterB(self.chatbot)
-        self.chatbot.system_logic_adapters = [adapter_a]
-        self.chatbot.logic_adapters = [adapter_b]
-
-        self.assertIsLength(self.chatbot.get_logic_adapters(), 2)
-        self.assertIn(adapter_a, self.chatbot.get_logic_adapters())
-        self.assertIn(adapter_b, self.chatbot.get_logic_adapters())
+        self.assertEqual(statement.text, 'Good morning.')
 
     def test_chatbot_set_for_all_logic_adapters(self):
-        for sub_adapter in self.chatbot.get_logic_adapters():
+        for sub_adapter in self.chatbot.logic_adapters:
             self.assertEqual(sub_adapter.chatbot, self.chatbot)
         self.assertGreater(
-            len(self.chatbot.get_logic_adapters()), 0,
+            len(self.chatbot.logic_adapters), 0,
             msg='At least one logic adapter is expected for this test.'
         )
 
