@@ -1,5 +1,4 @@
 from unittest import TestCase
-from chatterbot import __version__
 from chatterbot import __main__ as main
 
 
@@ -10,8 +9,7 @@ class CommandLineInterfaceTests(TestCase):
 
     def test_get_chatterbot_version(self):
         version = main.get_chatterbot_version()
-        self.assertEqual(version, __version__)
-
-    def test_get_nltk_data_directories(self):
-        directories = main.get_nltk_data_directories()
-        self.assertIn('/', directories)
+        version_parts = version.split('.')
+        self.assertEqual(len(version_parts), 3)
+        self.assertTrue(version_parts[0].isdigit())
+        self.assertTrue(version_parts[1].isdigit())
