@@ -1,5 +1,6 @@
 from tests.base_case import ChatBotTestCase
 from chatterbot.trainers import ListTrainer
+from chatterbot import preprocessors
 
 
 class ListTrainingTests(ChatBotTestCase):
@@ -10,6 +11,24 @@ class ListTrainingTests(ChatBotTestCase):
             self.chatbot,
             show_training_progress=False
         )
+
+    def test_training_cleans_whitespace(self):
+        """
+        Test that the ``clean_whitespace`` preprocessor is used during
+        the training process.
+        """
+        self.chatbot.preprocessors = [preprocessors.clean_whitespace]
+
+        self.trainer.train([
+            'Can I help you with anything?',
+            'No, I     think I am all set.',
+            'Okay, have a nice day.',
+            'Thank you, you too.'
+        ])
+
+        response = self.chatbot.get_response('Can I help you with anything?')
+
+        self.assertEqual(response.text, 'No, I think I am all set.')
 
     def test_training_adds_statements(self):
         """
@@ -65,7 +84,7 @@ class ListTrainingTests(ChatBotTestCase):
         ))
 
         self.assertIsLength(statements, 1)
-        self.assertEqual(statements[0].search_text, "ik")
+        self.assertEqual(statements[0].search_text, 'VERB:hat')
 
     def test_training_sets_search_in_response_to(self):
 
@@ -81,7 +100,7 @@ class ListTrainingTests(ChatBotTestCase):
         ))
 
         self.assertIsLength(statements, 1)
-        self.assertEqual(statements[0].search_in_response_to, "ik")
+        self.assertEqual(statements[0].search_in_response_to, 'VERB:hat')
 
     def test_database_has_correct_format(self):
         """
@@ -136,7 +155,7 @@ class ListTrainingTests(ChatBotTestCase):
 
         response = self.chatbot.get_response(conversation[1])
 
-        self.assertEqual(response, conversation[2])
+        self.assertEqual(response.text, conversation[2])
 
     def test_training_with_emoji_characters(self):
         """
@@ -152,7 +171,7 @@ class ListTrainingTests(ChatBotTestCase):
 
         response = self.chatbot.get_response(conversation[1])
 
-        self.assertEqual(response, conversation[2])
+        self.assertEqual(response.text, conversation[2])
 
     def test_training_with_unicode_bytestring(self):
         """
@@ -168,7 +187,7 @@ class ListTrainingTests(ChatBotTestCase):
 
         response = self.chatbot.get_response(conversation[1])
 
-        self.assertEqual(response, conversation[2])
+        self.assertEqual(response.text, conversation[2])
 
     def test_similar_sentence_gets_same_response_multiple_times(self):
         """
@@ -176,21 +195,31 @@ class ListTrainingTests(ChatBotTestCase):
         question (which is similar to the one present in the training set)
         when asked repeatedly.
         """
-        training = [
+        training_data = [
             'how do you login to gmail?',
             'Goto gmail.com, enter your login information and hit enter!'
         ]
 
         similar_question = 'how do I login to gmail?'
 
-        self.trainer.train(training)
+        self.trainer.train(training_data)
 
-        response_to_trained_set = self.chatbot.get_response('how do you login to gmail?')
-        response1 = self.chatbot.get_response(similar_question)
-        response2 = self.chatbot.get_response(similar_question)
+        response_to_trained_set = self.chatbot.get_response(
+            text='how do you login to gmail?',
+            conversation='a'
+        )
+        response1 = self.chatbot.get_response(
+            text=similar_question,
+            conversation='b'
+        )
+        response2 = self.chatbot.get_response(
+            text=similar_question,
+            conversation='c'
+        )
 
-        self.assertEqual(response_to_trained_set, response1)
-        self.assertEqual(response1.text, response2.text)
+        self.assertEqual(response_to_trained_set.text, training_data[1])
+        self.assertEqual(response1.text, training_data[1])
+        self.assertEqual(response2.text, training_data[1])
 
     def test_consecutive_trainings_same_responses_different_inputs(self):
         """
